// src/app/models/store.ts
export class Store {
    AverageScore:number;
    AverageShippingDelay:number;
    Id: number;
    ItemId: number;
    Logo: string;
    LookAndFeelId: number;
    PointsByPesoExchange:number;
    Removed:boolean;
    ShippingArea:number;
    URL: string;
    Telephone: string;
    Delivery:string;
    Mail: string;
    Name:string;
    Longitude: number;
    Latitude: number;
    Description:string;
}

export class NewStore {
    AverageScore:number;
    AverageShippingDelay:number;
    Id: number;
    ItemId: number;
    Logo: string;
    LookAndFeelId: number;
    PointsByPesoExchange:number;
    Removed:boolean;
    ShippingArea:number;
    URL: string;
    Telephone: string;
    Delivery:string;
    Mail: string;
    Name:string;
    Longitude: number;
    Latitude: number;
    Description:string;
    UserId:string;
}

// catergory todo lo de item